+++ tests/unit/helpers/parse-meme-url-test.ts
@@ -6,16 +6,16 @@ module('Unit | Helper | parse-meme-url', function (hooks) {
   setupTest(hooks);
 
   test('returns the input', function (assert) {
-    assert.equal(
+    assert.strictEqual(
       parseMemeUrl(['https://some-absolue-path.com']),
       'https://some-absolue-path.com'
     );
   });
 
   test('transforms a relative URL to the appropiate absolute URL', function (assert) {
-    assert.equal(
+    assert.strictEqual(
       parseMemeUrl(['some/meme.png']),
-      'https://test.spuxx.de/images/memes/some-meme.png'
+      'https://test.potber.de/images/memes/some/meme.png'
     );
   });
 });